Change amount value from int to long

+++ src/test/java/com/checkout/TestHelper.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 
 public class TestHelper {
     public static PaymentRequest<CardSource> createCardPaymentRequest() {
-        return createCardPaymentRequest(100);
+        return createCardPaymentRequest(100L);
     }
 
     public static PaymentRequest<CardSource> createCardPayoutRequest() {
@@ -17,7 +17,7 @@ public class TestHelper {
         cardSource.setLastName("Doe");
 
 
-        PaymentRequest<CardSource> request = PaymentRequest.fromDestination(cardSource, Currency.GBP, 100);
+        PaymentRequest<CardSource> request = PaymentRequest.fromDestination(cardSource, Currency.GBP, 100L);
         request.setReference(UUID.randomUUID().toString());
         Processing processing = Processing.builder().senderInformation(
                 SenderInformation.builder()
@@ -38,7 +38,7 @@ public class TestHelper {
         return request;
     }
 
-    public static PaymentRequest<CardSource> createCardPaymentRequest(Integer amount) {
+    public static PaymentRequest<CardSource> createCardPaymentRequest(Long amount) {
         CardSource cardSource = new CardSource(TestCardSource.VISA.getNumber(), TestCardSource.VISA.getExpiryMonth(), TestCardSource.VISA.getExpiryYear());
         cardSource.setCvv(TestCardSource.VISA.getCvv());
 
@@ -57,7 +57,7 @@ public class TestHelper {
         return createAlternativePaymentMethodRequest(alternativePaymentMethodRequestSource, currency, 100);
     }
 
-    public static PaymentRequest<RequestSource> createAlternativePaymentMethodRequest(RequestSource alternativePaymentMethodRequestSource, String currency, int amount) {
+    public static PaymentRequest<RequestSource> createAlternativePaymentMethodRequest(RequestSource alternativePaymentMethodRequestSource, String currency, long amount) {
         CustomerRequest customer = new CustomerRequest();
         customer.setEmail(generateRandomEmail());
 
@@ -80,7 +80,7 @@ public class TestHelper {
     }
 
     public static PaymentRequest<TokenSource> createTokenPaymentRequest(String token) {
-        PaymentRequest<TokenSource> request = PaymentRequest.fromSource(new TokenSource(token), Currency.GBP, 100);
+        PaymentRequest<TokenSource> request = PaymentRequest.fromSource(new TokenSource(token), Currency.GBP, 100L);
         request.setCapture(false);
         return request;
     }
